refactor: Asset / Add subfolders #254
