+++ client/src/components/Asset/Editor/AssetCardEditor.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const AssetCardEditor = () => {
+  return (
+    <div>AssetCardEditor</div>
+  )
+}
+
+export default AssetCardEditor
