+++ client/src/components/Asset/List/AssetList.tsx
@@ -1,6 +1,6 @@
-import Amount from '../../models/Amount';
-import { AssetProps } from '../../screens/Asset';
-import Icon from '../UI/Icon';
+import Amount from '../../../models/Amount';
+import { AssetProps } from '../../../screens/Asset';
+import Icon from '../../UI/Icon';
 import classes from './AssetList.module.css';
 import CardList from './CardList';
