+++ client/src/screens/Asset.tsx
@@ -1,5 +1,5 @@
-import AssetCardStatus from '../components/Asset/AssetCardStatus';
-import AssetList from '../components/Asset/AssetList';
+import AssetCardStatus from '../components/Asset/Status/AssetCardStatus';
+import AssetList from '../components/Asset/List/AssetList';
 import { useAppSelector } from '../hooks/redux-hook';
 import { AssetDataType, CardDataType } from '../util/api/assetAPI';
 import classes from './Asset.module.css';
